requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
black==25.1.0
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
ecdsa==0.19.1
fastapi==0.115.12
flake8==7.2.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycodestyle==2.13.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
google-cloud-firestore
google-auth
firebase-admin==6.5.0
